+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+ - Kit Editor: added internal versioning field to kits, that tells if they were
+   created using lsdpatcher 1.11.1 or above. this allows adding old kits to
+   lsdj 9.2.0+ without loss of quality.
 
 ## [1.11.0] - 2021-04-15
 ### Fixed